Added padding around text blocks
